Data type changed

+++ support_functions.py
@@ -58,14 +58,14 @@ def street_input_analysis(input):
                                 element[1]=element[1].replace(")", "")
                                 try:
                                     #number conversion
-                                    number_x = int((element[0][:]))
-                                    number_y = int((element[1][:]))
+                                    number_x = float((element[0][:]))
+                                    number_y = float((element[1][:]))
                                 except ValueError:
                                     #ensure all input is number
                                     raise ParseException("Please check the input of street coordinate")
                                 #add the result into vertice list
-                                vertex.append(int(number_x))
-                                vertex.append(int(number_y))
+                                vertex.append(float(number_x))
+                                vertex.append(float(number_y))
                                 vertices.append(vertex)
                         else:
                             raise ParseException("Please check the input of street coordinate")
